0.10.0

+++ scripts/init.d/mw.tpl
@@ -243,8 +243,16 @@ case "$1" in
             elif [ "$v4" != "" ]; then
                 address="MW-Panel-Url: http://$v4:$port$auth_path"
             elif [ "$v6" != "" ]; then
-                echo 'True' > $mw_path/data/ipv6.pl
+
+                if [ ! -f $mw_path/data/ipv6.pl ];then
+                    #  Need to restart ipv6 to take effect
+                    mw_stop
+                    mw_start
+                    echo 'True' > $mw_path/data/ipv6.pl
+                fi
+                
                 address="MW-Panel-Url: http://[$v6]:$port$auth_path"
+
             else
                 address="MW-Panel-Url: http://you-network-ip:$port$auth_path"
             fi
